server: fix the server not starting error
1. upgrade glassfish to version 5.0
vesrion 4.1.1 has a problem with creating new resources. the version 5.0
comes to solve it

2. sercurity/BCrypt.java
include an encryption algorithm to encrypt the password

3. entities.*
fix the problem with the foreignkey reference in child tables. add
custom names to all fiels that were conflicting with the in-built names
in derby

4. authentication/Register.java
add sample code for registration(failed)

5. README.md
give more details on how to run the project